Update the prepare command to gather US files
Updates the logic of the `prepare` command to gather the files required
for the US in case of an american city.

These files were previously ignored since the BNA has some specific
logic to fetch the for American cities. However, in order to make the
BNA more generic, it makes sense to enhance this command to make it
exhaustive.

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>