Add Dev container feature for local development
Uses a single multistage Dockerfile to create a dev container for local
development. Several packages needed for development, especially for
running the Just recipes, that are not included in the base image are
installed for the dev container. The documentation is also updated with
instructions on how to use the dev container.