Migrate to uv (#817)
Migrates from `poetry` to `uv`.

We had been wanting to do so for a while and now that `poetry export` is
causing us some trouble, this is a good excuse.

The `justfile` and the `Dockerfile`, the shared GitHub workflows, the
project workflows and the documentation have been adjusted to use `uv`
as well.

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>