Use cache folder for downloads
Uses platform specific cache folder to download files common to all
analyses.

The cache can also be an AWS S3 bucket.

When fetching file, the downloader always looks for the file in the
cache first, then either fetches it from the internet if it does not
exist, or simply copy it from the cache to the data store.

Drive-bys:
- Update dependencies

Fixes PeopleForBikes/brokenspoke-analyzer#254

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>