Use cache folder for downloads (#842)
Uses platform specific cache folder to download files common to all
analyses.

The cache can also be an AWS S3 bucket.

When fetching file, the downloader always looks for the file in the
cache first, then either fetches it from the internet if it does not
exist, or simply copy it from the cache to the data store.

The documentation was updated accordingly.

Since the cache really speeds up the "prepare" stage, the operations
faster than 80ms were not being displayed on screen as the `Console`
object did not have time to refresh. As a result, the output was
adjusted to always display log messages, but only show the spinner if
the operation takes longer than 80 ms, which seemd to be a reasonable
compromise.

Drive-bys:
- Updates runtime and dev dependencies.
- Adds a script to warmup the cache.
- Fixes SQL lint violations.

Fixes PeopleForBikes/brokenspoke-analyzer#254

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>