Add ability to run partial analysis (#845)
Adds the ability to select wich parts of the analysis to run (e.g.
features, stress or connectivity).

The compute logic was updated to account for the dependencies between
the different parts. For instance, the "stress" part requires the
"features" part to run. This determination is made automatically for the
user.

The export logic was also updated to prevent non-existing tables to be
exporting, which would result in a crash. Indeed, if not all the parts
were run, not all the tables were created.

The documentation was updated accordingly.

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>