Fix mileage calculation
Before, mileage was calculated and reported in a directional way, so
there was a mileage for each of ft_ and tf_. This commit changes
the calculation so that mileage is added up for both directions.
A LATERAL join is used to avoid scanning the same table twice.

Drive-bys:
- Fix the compute command to actually only run the partial parts.

Fixes PeopleForBikes/brokenspoke-analyzer#858