Fix postgres healthcheck command (#914)
Fix postgres healthcheck error by adding a user to the command

When running the healthcheck command, postgres was trying to connect
using the system user, which in a container shell is root,
so we were seeing the message "FATAL:  role "root" does not exist".
Adding a user to the command fixes this issue.

Fixes: PeopleForBikes/brokenspoke-analyzer#898