BNA cache (#894)
Add bna cache subcommand

Adds the `bna cache` subcommand and updates the caching mechanism
according to the new behavior discussed for v3.

The `bna prepare`, `bna run` and `bna run-with` commands were updated to
handle the cache and mirror flags.

This PR also makes sure to use 2020 Census Population and Employment
data in the `datastore` module. The `downloader` module was removed, as
well as all the references to it.

A new `file_utils` module was created to handle the types and functions
performing file system related operations. Doctests were added when
appropriate.

The cache was disabled in the CI.

Signed-off-by: Rémy Greinhofer <remy.greinhofer@gmail.com>
Co-authored-by: Mitchell Henke <mitchell@mitchellhenke.com>