Improve the usage of the dev container (#925)
* Add a VS Code debugger configuration to run the bna

Also track the debugger configuration in git.

* Add a function to reset the database for a new run

When running the brokenspoke-analyzer in a dev container, it is
useful to be able to reset the database tables created by a
previous run since making a new run requires a clean database.
This omits the need to rebuild the entire postgres container
for a new run.

* Add a CLI command for the reset database tables functionality

The command looks like this:
bna configure reset <database_url> <pguser>